+++ docs/release-notes.rst
@@ -1,6 +1,13 @@
 Release notes
 =============
 
+3.0.7
+------
+
+    - File extension will NOT be added automatically for output filenames. User specified filename is uses as it is.
+    - Documentation improved. Parameter type hint added to every parameter
+    - Chemical shift values in API dump are now converted into float before processing.
+
 3.0.6
 ------